Refactor/#5: 친구요청 API의 요청DTO 와 응답DTO 네이밍 및 위치 변경
1. 친구요청 API에서 사용되는 요청 DTO 위치 변경
2. 친구요청 API에서 사용되는 응답 DTO 이름 변경 - SendFriendsResponseDto
