+++ src/main/java/com/sns/api/friends/controller/FriendsController.java
@@ -1,7 +1,8 @@
 package com.sns.api.friends.controller;
 
 import com.sns.api.common.domain.dto.UserBaseDto;
-import com.sns.api.friends.domain.dto.response.FriendsResponseDto;
+import com.sns.api.friends.domain.dto.request.SendFriendsRequestDto;
+import com.sns.api.friends.domain.dto.response.SendFriendsResponseDto;
 import com.sns.api.friends.service.FriendsService;
 import com.sns.common.component.BaseResponse;
 import com.sns.common.component.Const;
@@ -18,8 +19,8 @@ public class FriendsController {
     private final FriendsService friendsService;
 
     @PostMapping("/requests")
-    public BaseResponse<FriendsResponseDto> requestFriend(@RequestBody @Valid SendFriendsRequestDto requestDto,
-                                                          @SessionAttribute(Const.LOGIN_USER) UserBaseDto userBaseDto) {
+    public BaseResponse<SendFriendsResponseDto> requestFriend(@RequestBody @Valid SendFriendsRequestDto requestDto,
+                                                              @SessionAttribute(Const.LOGIN_USER) UserBaseDto userBaseDto) {
         return BaseResponse.success(friendsService.requestFriend(requestDto, userBaseDto), ResultCode.CREATED);
     }
 }
