--- src/main/java/com/sns/api/friends/domain/dto/request/FriendsRequestDto.java
@@ -1,12 +1,0 @@
-package com.sns.api.friends.domain.dto.request;
-
-import jakarta.validation.constraints.NotNull;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class FriendsRequestDto {
-    @NotNull(message = "receiverId는 필수입니다.")
-    private final Long receiverId;
-}
